Feat: CallManager Singleton으로 수정
